Ifpack: changes made to support 64-bit EpetraExt and Ifpack. Code works in multiple modes (32 only, 64 only, 32 and 64).
Build/Test Cases Summary
Enabled Packages: Epetra, Teuchos
Disabled Packages: SEACAS,Sundance,STK
Enabled all Forward Packages
0) MPI_DEBUG => FAILED: passed=1729,notpassed=64 => Not ready to push! (48.64 min)
1) SERIAL_RELEASE => passed: passed=1700,notpassed=0 (62.70 min)
Other local commits for this build/test group: 0571311, da0b89d, c3d832e, 47135cd
WARNING: Forced the push!

+++ packages/ifpack/src/Ifpack_Krylov.h
@@ -341,9 +341,9 @@ private:
   //! Number of local nonzeros.
   int NumMyNonzeros_;
   //! Number of global rows.
-  int NumGlobalRows_;
+  long long NumGlobalRows_;
   //! Number of global nonzeros.
-  int NumGlobalNonzeros_;
+  long long NumGlobalNonzeros_;
   //! Pointers to the matrix as an Epetra_Operator.
   Teuchos::RefCountPtr<Epetra_Operator> Operator_;
   //! Pointers to the matrix as an Epetra_RowMatrix.
